fix: expose all overlay items
